+++ .gitignore
@@ -1,1 +1,2 @@
 *~
+/CMakeLists.txt.user
