+++ gtest/gtest_tree.cpp
@@ -14,7 +14,7 @@
 #include <gtest/gtest.h>
 #include "action_test_node.h"
 #include "condition_test_node.h"
-#include "behavior_tree/behavior_tree.h"
+#include "behavior_tree_core/behavior_tree.h"
