--- include/behavior_tree/behavior_tree.h
@@ -1,20 +1,0 @@
-#ifndef BEHAVIOR_TREE_H
-#define BEHAVIOR_TREE_H
-
-#include "behavior_tree/parallel_node.h"
-#include "behavior_tree/fallback_node.h"
-#include "behavior_tree/sequence_node.h"
-
-#include "behavior_tree/sequence_node_with_memory.h"
-#include "behavior_tree/fallback_node_with_memory.h"
-
-#include "behavior_tree/action_node.h"
-#include "behavior_tree/condition_node.h"
-
-#include "behavior_tree/decorator_negation_node.h"
-#include "behavior_tree/decorator_retry_node.h"
-
-#include "behavior_tree/exceptions.h"
-
-
-#endif // BEHAVIOR_TREE_H
