+++ include/behavior_tree_core/behavior_tree.h
@@ -1,0 +1,20 @@
+#ifndef BEHAVIOR_TREE_H
+#define BEHAVIOR_TREE_H
+
+#include "behavior_tree_core/parallel_node.h"
+#include "behavior_tree_core/fallback_node.h"
+#include "behavior_tree_core/sequence_node.h"
+
+#include "behavior_tree_core/sequence_node_with_memory.h"
+#include "behavior_tree_core/fallback_node_with_memory.h"
+
+#include "behavior_tree_core/action_node.h"
+#include "behavior_tree_core/condition_node.h"
+
+#include "behavior_tree_core/decorator_negation_node.h"
+#include "behavior_tree_core/decorator_retry_node.h"
+
+#include "behavior_tree_core/exceptions.h"
+
+
+#endif // BEHAVIOR_TREE_H
