+++ include/behavior_tree_core/condition_node.h
