+++ include/behavior_tree_core/decorator_negation_node.h
@@ -1,7 +1,7 @@
 #ifndef DECORATORNEGATIONNODE_H
 #define DECORATORNEGATIONNODE_H
 
-#include "behavior_tree/control_node.h"
+#include "behavior_tree_core/control_node.h"
 
 namespace BT
 {
