+++ include/behavior_tree_core/exceptions.h
