+++ include/behavior_tree_core/fallback_node.h
@@ -1,7 +1,7 @@
 #ifndef FALLBACKNODE_H
 #define FALLBACKNODE_H
 
-#include "behavior_tree/control_node.h"
+#include "behavior_tree_core/control_node.h"
 
 namespace BT
 {
