+++ include/behavior_tree_core/fallback_node_with_memory.h
@@ -1,6 +1,6 @@
 #ifndef FALLBACK_NODE_WITH_MEMORY_H
 #define FALLBACK_NODE_WITH_MEMORY_H
-#include "behavior_tree/control_node.h"
+#include "behavior_tree_core/control_node.h"
 
 
 namespace BT
