+++ include/behavior_tree_core/sequence_node.h
@@ -1,7 +1,7 @@
 #ifndef SEQUENCENODE_H
 #define SEQUENCENODE_H
 
-#include "behavior_tree/control_node.h"
+#include "behavior_tree_core/control_node.h"
 
 namespace BT
 {
