+++ include/behavior_tree_core/sequence_node_with_memory.h
@@ -2,7 +2,7 @@
 #define SEQUENCE_NODE_WITH_MEMORY_H
 
 
-#include "behavior_tree/control_node.h"
+#include "behavior_tree_core/control_node.h"
