+++ src/condition_node.cpp
@@ -10,7 +10,7 @@
 *   WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-#include "behavior_tree/condition_node.h"
+#include "behavior_tree_core/condition_node.h"
 #include <string>
 
 BT::ConditionNode::ConditionNode(std::string name) :
