+++ src/exceptions.cpp
@@ -10,7 +10,7 @@
 *   WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-#include "behavior_tree/exceptions.h"
+#include "behavior_tree_core/exceptions.h"
 #include<string>
 
 BT::BehaviorTreeException::BehaviorTreeException(const std::string Message)
