+++ src/fallback_node.cpp
@@ -10,7 +10,7 @@
 *   WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-#include "behavior_tree/fallback_node.h"
+#include "behavior_tree_core/fallback_node.h"
 #include <string>
 
 BT::FallbackNode::FallbackNode(std::string name) : ControlNode::ControlNode(name) {}
