+++ src/sequence_node.cpp
@@ -12,7 +12,7 @@
 */
 
 
-#include "behavior_tree/sequence_node.h"
+#include "behavior_tree_core/sequence_node.h"
 #include <string>
