+++ src/tick_engine.cpp
@@ -10,7 +10,7 @@
 *   WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-#include "behavior_tree/tick_engine.h"
+#include "behavior_tree_core/tick_engine.h"
 
 TickEngine::TickEngine(int initial_value)
 {
